Updated the issues from Rigol_Dp932U and Unit test file
